+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## Unreleased
+
+- First K8s.Client.Runner.Stream evaluation made lazy
+- K8s.Resource.api_version/1
+
 ## [0.2.11] - 2019-06-24
 
 ### Added